requirements.txt
# CGR Assessment - Unified Container Vulnerability Assessment Tool
# Requirements for both HTML and XLSX output formats

# Core dependencies
docker>=7.0.0
xlsxwriter>=3.2.0
requests>=2.32.0
pyyaml>=6.0.0

# Optional: Enhanced markdown support for HTML reports
markdown>=3.0.0

# Development dependencies (testing)
pytest>=8.0.0
pytest-cov>=4.1.0
